+++ main/dzVents.cpp
@@ -25,7 +25,7 @@ extern http::server::CWebServerHelper m_webservers;
 CdzVents CdzVents::m_dzvents;
 
 CdzVents::CdzVents(void) :
-	m_version("3.0.13")
+	m_version("3.0.14")
 {
 	m_bdzVentsExist = false;
 }