+++ dzVents/documentation/history.md
@@ -1,7 +1,10 @@
+[3.0.14]
+- Add utils.fuzzyLookup
+- Made eventHelpers more resilient to coding errors in the on = section
+
 [3.0.13]
 - Add utils.osCommand (caption)
 - Improved utils.stringSplit (implicit conversion to number on request)
-- Made eventHelpers more resilient to coding errors in the on = section
 - Made log function more resilient when logitem is table with embedded function(s)
 - Improved domoticz.snapshot() to enable snapshot of multiple cameras (in one Email)